api/requirements.txt
# Azure Functions requirements
azure-functions
azure-cosmos